Move _get_free_symbols to _operations
